OSIS-791 Rename files in according with CRUD
