+++ base/templates/learning_unit/blocks/tabs.html
