+++ base/templates/learning_unit/blocks/type.html
@@ -1,4 +1,3 @@
-
 {% load bootstrap3 %}
 {% load i18n %}
