+++ base/templates/learning_unit/by_activity.html
@@ -1,4 +1,3 @@
-
 {% load staticfiles %}
 {% load i18n %}
 {% load learning_unit %}
@@ -70,7 +69,7 @@
                     <td>{{ learning_unit.complete_title|default_if_none:'' }}</td>
                     <td>
                         {% with learning_unit.learning_container_year.container_type|default_if_none:'-' as type %}
-                        {% trans type %}
+                       submit {% trans type %}
                         {% endwith %}
                     </td>
                     <td>
