+++ base/templates/learning_unit/components.html
@@ -35,10 +35,10 @@
 {% block content %}
 {% bootstrap_messages %}
 
-{% include "learning_unit/header.html" %}
+{% include "learning_unit/blocks/header.html" %}
 <div class="panel panel-default">
     <div class="panel-body" id="tabs">
-        {% include "learning_unit/tabs.html" %}
+        {% include "learning_unit/blocks/tabs.html" %}
         <div class="tab-content" id="tab_content">
             <div role="tabpanel" class="tab-pane active" id="components">
                 <table class="table">
