+++ base/templates/learning_unit/components/volumes_management_form.html
@@ -77,5 +77,5 @@
         </a>
         <button type="button" class="btn btn-primary" data-toggle="modal" data-target="#modalReport">{% trans 'save' %}</button>
     </div>
-    {% include "learning_unit/components/modal_modification_ask_to_report.html" %}
+    {% include "learning_unit/blocks/modal/modal_modification_ask_to_report.html" %}
 </form>
