+++ base/templates/learning_unit/confirm_delete.html
