+++ base/templates/learning_unit/identification.html
@@ -33,7 +33,7 @@
 
 {% block content %}
 
-{% include "learning_unit/header.html" %}
+{% include "learning_unit/blocks/header.html" %}
 {% include "template_messages.html" %}
 
 <div class="panel panel-default">
@@ -41,7 +41,7 @@
         {% include "learning_unit/identification_actions_bar.html" with dropdirection="dropdown"%}
     </div>
     <div class="panel-body" id="tabs" >
-        {% include "learning_unit/tabs.html" %}
+        {% include "learning_unit/blocks/tabs.html" %}
         <div class="tab-content">
             <div role="tabpanel" class="tab-pane active" id="identification">
                 <br/>
