+++ base/templates/learning_unit/learning_unit_layout.html
@@ -19,7 +19,7 @@
 {% block content %}
 
 {% with disabled_nav=True %}
-  {% include "learning_unit/header.html" %}
+  {% include "learning_unit/blocks/header.html" %}
 {% endwith %}
 
 <div class="panel panel-default" >
