+++ base/templates/learning_unit/modification.html
@@ -28,5 +28,6 @@
 {% endblock submit %}
 
 {% block modal %}
-    {% include "learning_unit/components/modal_modification_ask_to_report.html" %}
+    {% include "learning_unit/blocks/modal/modal_modification_ask_to_report.html" %}
+    {% include "learning_unit/blocks/modal/modal_confirm_postement.html" %}
 {% endblock modal %}
