+++ base/templates/learning_unit/pedagogy.html
@@ -31,10 +31,10 @@
 {% load getattr %}
 
 {% block content %}
-{% include "learning_unit/header.html" %}
+{% include "learning_unit/blocks/header.html" %}
 <div class="panel panel-default" >
     <div class="panel-body" id="tabs" >
-        {% include "learning_unit/tabs.html" %}
+        {% include "learning_unit/blocks/tabs.html" %}
         <div class="tab-content" id="tab_content">
             <div role="tabpanel" class="tab-pane active" id="pedagogy">
                  {% url 'learning_unit_pedagogy_edit' learning_unit_year_id=learning_unit_year.id as post_url%}
