+++ base/templates/learning_unit/specifications.html
@@ -33,10 +33,10 @@
 
 {% block content %}
 
-{% include "learning_unit/header.html" %}
+{% include "learning_unit/blocks/header.html" %}
 <div class="panel panel-default" >
     <div class="panel-body" id="tabs" >
-        {% include "learning_unit/tabs.html" %}
+        {% include "learning_unit/blocks/tabs.html" %}
 
         <div class="tab-content" id="tab_content">
             <div role="tabpanel" class="tab-pane active" id="specifications">
