+++ base/tests/views/test_learning_unit_deletion.py
@@ -101,7 +101,7 @@ class LearningUnitDelete(TestCase):
     def test_delete_from_given_learning_unit_year_case_success(self, mock_render):
         learning_unit_years = self.learning_unit_year_list
 
-        from base.views.learning_unit_deletion import delete_from_given_learning_unit_year
+        from base.views.learning_units.delete import delete_from_given_learning_unit_year
 
         request_factory = RequestFactory()
 
@@ -142,7 +142,7 @@ class LearningUnitDelete(TestCase):
     def test_delete_all_learning_units_year_method_not_allowed(self):
         learning_unit_years = self.learning_unit_year_list
 
-        from base.views.learning_unit_deletion import delete_all_learning_units_year
+        from base.views.learning_units.delete import delete_all_learning_units_year
 
         request_factory = RequestFactory()
         request = request_factory.get(reverse(delete_all_learning_units_year, args=[learning_unit_years[1].id]))
@@ -154,7 +154,7 @@ class LearningUnitDelete(TestCase):
     def test_delete_all_learning_units_year_case_success(self):
         learning_unit_years = self.learning_unit_year_list
 
-        from base.views.learning_unit_deletion import delete_all_learning_units_year
+        from base.views.learning_units.delete import delete_all_learning_units_year
 
         request_factory = RequestFactory()
 
@@ -183,7 +183,7 @@ class LearningUnitDelete(TestCase):
         ly1 = learning_unit_years[1]
         LearningUnitEnrollmentFactory(learning_unit_year=ly1)
 
-        from base.views.learning_unit_deletion import delete_from_given_learning_unit_year
+        from base.views.learning_units.delete import delete_from_given_learning_unit_year
 
         request_factory = RequestFactory()
 
@@ -218,7 +218,7 @@ class LearningUnitDelete(TestCase):
         ly1 = learning_unit_years[1]
         LearningUnitEnrollmentFactory(learning_unit_year=ly1)
 
-        from base.views.learning_unit_deletion import delete_all_learning_units_year
+        from base.views.learning_units.delete import delete_all_learning_units_year
 
         request_factory = RequestFactory()
 
@@ -258,7 +258,7 @@ class LearningUnitDelete(TestCase):
         ly1 = learning_unit_years[1]
         LearningUnitEnrollmentFactory(learning_unit_year=ly1)
 
-        from base.views.learning_unit_deletion import delete_from_given_learning_unit_year
+        from base.views.learning_units.delete import delete_from_given_learning_unit_year
 
         request_factory = RequestFactory()
 
@@ -289,7 +289,7 @@ class LearningUnitDelete(TestCase):
         self.assertIsNotNone(LearningUnitYear.objects.get(id=ly1.id))
 
     def test_delete_from_given_learning_unit_year_case_error_not_attached_to_entity(self):
-        from base.views.learning_unit_deletion import delete_from_given_learning_unit_year
+        from base.views.learning_units.delete import delete_from_given_learning_unit_year
 
         l_unit_year_to_delete = self.learning_unit_year_list[0]
         PersonEntity.objects.all().delete()  # Remove all link person entity
@@ -303,7 +303,7 @@ class LearningUnitDelete(TestCase):
 
     def test_delete_from_given_learning_unit_year_faculty_manager_role(self):
         """A Faculty manager can only remove container_type other than COURSE/INTERNSHIP/DISSERTATION"""
-        from base.views.learning_unit_deletion import delete_from_given_learning_unit_year
+        from base.views.learning_units.delete import delete_from_given_learning_unit_year
 
         add_to_group(self.user, person.FACULTY_MANAGER_GROUP)
         l_unit_year_to_delete = self.learning_unit_year_list[0]
@@ -324,7 +324,7 @@ class LearningUnitDelete(TestCase):
 
     def test_delete_from_given_learning_unit_year_case_error_faculty_manager(self):
         """A Faculty manager can only remove container_type other than COURSE/INTERNSHIP/DISSERTATION"""
-        from base.views.learning_unit_deletion import delete_from_given_learning_unit_year
+        from base.views.learning_units.delete import delete_from_given_learning_unit_year
 
         add_to_group(self.user, person.FACULTY_MANAGER_GROUP)
         l_unit_year_to_delete = self.learning_unit_year_list[0]
